Distance to Distro Centers Test
Adding the notebook I used to test the distance to distribution centers calculation, alongside the janky table of zip codes that I used in the process. Yes, I know it's a text file, but it's a table. I dunno, ask the guy I got this from why they format their data like that.